Move movePiece into helper function
