+++ src/components/FreeMode/FreeMode.tsx
@@ -6,15 +6,7 @@ import MovesList from '../MovesList'
 import CommandColumn from '../CommandColumn'
 import styles from "./FreeMode.scss"
 import { PieceName, Move } from "../../sharedTypes"
-
-
-const pieceNameToNotation = {
-  "king": "K",
-  "queen": "Q",
-  "rook": "R",
-  "bishop": "B",
-  "knight": "N",
-}
+import movePiece from "../../helpers/movePiece"
 
 const FreeMode = () => { 
   const gameClient = chess.create()
@@ -31,63 +23,6 @@ const FreeMode = () => {
         [...prevMoves.slice(0, prevMoves.length - 1), [lastTurn[0], {notation: move}]] :
         [...prevMoves, [{notation: move}]]
     });
-  }
-
-  const movePiece = (
-    piece: PieceName,
-    targetSquare: string,
-    currentLocation: {rank: number, file: string},
-    isTaking: boolean
-  ) => {
-    const { rank, file } = currentLocation
-    targetSquare = isTaking ? "x" + targetSquare : targetSquare
-    try {
-      if ( piece === "pawn" ) {
-        const move = isTaking ?
-          file + targetSquare :
-          targetSquare
-        gameClient.move(move)
-        addMoveToList(move)
-      } else if ( piece === "king" ) {
-        const moveNotation = "K" + targetSquare
-        try { 
-          gameClient.move(moveNotation) 
-          addMoveToList(moveNotation)
-        } catch {
-          if (targetSquare === "g1" || targetSquare === "g8") {
-            gameClient.move("0-0")
-            addMoveToList("0-0")
-          } else if (targetSquare === "c1" || targetSquare === "c8") {
-            gameClient.move("0-0-0")
-            addMoveToList("0-0-0")
-          } else {
-            throw Error("Invalid King move")
-          }
-        }
-      } else {
-        const pieceNotation = pieceNameToNotation[piece]
-        const moveNotation = pieceNotation + targetSquare
-        try {
-          gameClient.move(moveNotation)
-          addMoveToList(moveNotation)
-        // Passing rank & file together isn't supported (eg. Nb1C3)
-        } catch(e) {
-          try {
-            const move = pieceNotation + file + targetSquare
-            gameClient.move(move)
-            addMoveToList(move)
-          } catch(e) {
-            const move = pieceNotation + rank + targetSquare
-            gameClient.move(move)
-            addMoveToList(move)
-          }
-        }
-      }
-      setCommandColumnMessage('')
-    } catch {
-      setCommandColumnMessage("Invalid move")
-    }
-    setBoard({ ...gameClient.game.board })
   }
 
   const addReasonToMove = () => {
@@ -109,7 +44,19 @@ const FreeMode = () => {
           <div className={styles.sideColumn}>
             <MovesList turns={moves}></MovesList>
           </div>
-          <Board movePiece={movePiece} board={board} isWhite={isWhite}></Board>
+          <Board 
+            movePiece={movePiece(
+              gameClient,
+              (moveNotation: string) => {
+                gameClient.move(moveNotation),
+                addMoveToList(moveNotation)
+              },
+              setCommandColumnMessage,
+              setBoard
+            )}
+            board={board}
+            isWhite={isWhite}
+          ></Board>
           <div className={styles.sideColumn}>
             <CommandColumn 
               setReasonInput={setReasonInput}
